refactor: isolate bcrypt from edge runtime
