+++ lib/auth/password.ts
@@ -1,0 +1,19 @@
+import bcrypt from 'bcryptjs'
+
+/**
+ * Hash a password using bcrypt.
+ */
+export async function hashPassword(password: string): Promise<string> {
+	const saltRounds = 12
+	return bcrypt.hash(password, saltRounds)
+}
+
+/**
+ * Verify a password against a hash.
+ */
+export async function comparePasswords(
+	password: string,
+	hash: string
+): Promise<boolean> {
+	return bcrypt.compare(password, hash)
+}
