+++ lib/utils.ts
@@ -1,4 +1,3 @@
-import bcrypt from 'bcryptjs'
 import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 import {
@@ -13,24 +12,6 @@ import {
 
 export function cn(...inputs: ClassValue[]) {
 	return twMerge(clsx(inputs))
-}
-
-/**
- * Hash a password using bcrypt
- */
-export async function hashPassword(password: string): Promise<string> {
-	const saltRounds = 12
-	return bcrypt.hash(password, saltRounds)
-}
-
-/**
- * Verify a password against a hash
- */
-export async function comparePasswords(
-	password: string,
-	hash: string
-): Promise<boolean> {
-	return bcrypt.compare(password, hash)
 }
 
 /**
