+++ server/actions/auth.ts
@@ -5,9 +5,9 @@ import { redirect } from 'next/navigation'
 import { AuthError } from 'next-auth'
 import { z } from 'zod'
 import { signIn, signOut } from '@/lib/auth'
+import { comparePasswords } from '@/lib/auth/password'
 import { CookieNames, getAvatarURL, Routes } from '@/lib/config'
 import { setEncryptedCookie } from '@/lib/cookies'
-import { comparePasswords } from '@/lib/utils'
 import { getAPI } from '@/server/api'
 import { AuthActionErrorCodeMap } from './constants'
 import { AuthErrorCodes, type ServerActionResponse } from './types'
